+++ sentry-apollo-3/src/main/java/io/sentry/apollo3/SentryApollo3HttpInterceptor.kt
@@ -223,6 +223,7 @@ constructor(
             span.spanContext.sampled = false
           }
         } catch (e: Throwable) {
+          span.spanContext.sampled = false
           scopes.options.logger.log(
             SentryLevel.ERROR,
             "An error occurred while executing beforeSpan on ApolloInterceptor",